accept both bytes & text to HostnameEndpoint since tests did both
also move idna encode/decode to new module
